Add pagination to the product listing in Bridge module
The commit introduces a new route /products that lists the products with pagination. Products are taken from the newly imported BridgeProductEntity and displayed with the help of the Flask's request module. Pagination parameters, 'page' and 'per_page', are received from the query parameters of the request. They are used to retrieve the corresponding page of products from the database and display the desired number of products per page.