Refactor categories.html to add image column
The update in categories.html introduces an additional table column for 'Images'. This new column lists the file name and type of each image associated with each category in the defined languages. This was added to provide a clear display of all images connected with a product category. The code for retrieving and displaying name and description has also been updated to directly access translated versions.