Enhance media handling in ERPArtikelController
Modified the logic for tax and category relation in ERPArtikelController.py. With new changes, if a tax or category does not exist in the database, a syncing operation will be attempted and afterwards, the tax or category is retrieved again from the database. Also updated the logic for setting categories by first clearing existing categories to avoid duplicates. This makes the process more reliable and efficient.