Add media handling to ERPArtikelEntity
Introduced functionality to handle media files in ERPArtikelEntity.py. Added an import for the BridgeMediaEntity and implemented its use in the ERPArtikelEntity class. Now media file information (name, type, size, title, description) can be assigned to the product entity. This allows for a better organization and handling of media files associated with products.