Extend category entity to include media attachments
Added functionality to include media attachments in the ERPArtikelKategorienEntity. This change was implemented using a newly added config class, ERPConfig. Now, for each ERP Category, image entities are created, matched with their metadata and attached. This was necessary to enable the ERP system to handle media files associated with category entities, improving the system's flexibility. Next, the language specification for the category translation changed from a constant 'DE_de' to a value derived from the category_nr_path. This helps in supporting more languages.