Add new progress bar HTML template
This commit introduces a new HTML template, `_progress.html`. This file contains the code for a progress bar designed to display a visual representation of a numerical value, with the progress being dynamically controlled via Vue.js. The progress bar will only display if the progress value is greater than or equal to zero.