Add SW6MarketplaceController for bridge entity management
Introduced a new controller file named SW6MarketplaceController.py to manage operational activities related to BridgeMarketplaceEntity in the SW6 module. This controller handles queries for specific BridgeMarketplaceEntities in the database as well as setting relationships between various entities. Its addition will streamline the management of entities in the SW6 Marketplace.