Modify accordion_item.html to enhance order details display
Implemented modifications to the accordion_item.html template to improve the display of order details. This includes checking if billing and shipping addresses are the same, displaying an "open" state indicator for orders, and adding a button to create orders. Also, fixed the file path for order_details.html and added the marketplace name to order information.