Add get_all functionality to BridgeAbstractController
A new method 'get_all' was added to the BridgeAbstractController making it possible to retrieve all entities in the database. The method includes error handling for potential exceptions that may be thrown during execution.

Even Darth Vader wouldn't argue the usefulness of this for inventory checks on the Death Star.

Signed-off-by: contact@get-company.com <contact@get-company.com>