Add UUID generation for BridgeAbstractEntity
A new method `set_sw6_id` was included for the `BridgeAbstractEntity` class. This method generates and returns a unique, random ID using `uuid4` from the `uuid` library, which is now being imported at the beginning of the file.

If only Darth Vader had thought to implement a UUID for each Stormtrooper, he could have easily tracked the incompetence that led to the fall of the Death Star. But alas, he overlooked this, and we all know how that catastrophe ended.

Signed-off-by: contact@get-company.com <contact@get-company.com>