Add search bar to category template
A search bar was introduced in the 'category' template. This change was applied to facilitate user's search within categories at a specific website.

It's like Darth Vader forgetting to install search functionality in the Death Star's intranet. Because who would need to find anything in a space station the size of a moon, right?

Signed-off-by: contact@get-company.com <contact@get-company.com>