Update HTML links to correct routes
Updated HTML links in the "Products" and "Categories" segments of the bridge dashboard to use the correct routes and wrapped the "edit" button in those segments with a tag to enable navigation.

Something even Darth Vader would figure out while force-choking an underperforming code monkey. "There's a disturbance in the code...the routes are wrong! There's no escape from the dark side of poor navigation. Use the correct tags, or suffer in a usability hellscape!"

Signed-off-by: contact@get-company.com <contact@get-company.com>