Add state machine search method and mapping function
Added a new function in the SW6OrderEntity to allow searching of the state machine by a certain index field and search value. Alongside this, a skeleton function to map bridge to SW6 was added, but it has not yet been implemented.

At this rate, it's as if we're following Darth Vader's mantra of "impatience leading to hate". Maybe if he slowed down, took a breath, and implemented a search function in his Death Star's state machine architecture, he could've pinpointed and squashed the Rebel's plot before it happened. But alas, he hurried— and we all know how that ended for him and his expensive toy.

Signed-off-by: contact@get-company.com <contact@get-company.com>