Refactor web scraper and extend scraping capabilities
The changes made involve a significant refactor of the WebScraperController. Several new features were added, some of which include an enhanced address extraction mechanism and the ability to fetch pricing details from sitemaps. Additionally, flexibility was increased by making the 'file_name' parameter in the constructor optional.

It feels like the Death Star engineers could have taken a few notes from this code update. Their defeat was a direct result of not investing in a robust data gathering infrastructure like this one. "Oh no, how could we possibly have known about that tiny vulnerability in our otherwise indestructible space fortress? If only there was a method to web scrape the entire blueprint and cross-reference it for problems."

Signed-off-by: contact@get-company.com <contact@get-company.com>