Update sidebar menu icons and add new sections
Changed the icons for the sidebar menu options and introduced new sections for enhanced navigation. This adds more structure and clarity to the interface, assisting the user in efficiently finding necessary options.

If Darth Vader were to do this change, he would have probably stated "The ability to alter a sidebar is insignificant next to the power of the Force". A shame he was too busy Force-choking his subordinates rather than focusing on user experience, that might have been the true downfall of the Empire.

Signed-off-by: contact@get-company.com <contact@get-company.com>